+++ modules/bridge/src/main/ts/ephox/bridge/components/dialog/ColorInput.ts
@@ -1,17 +1,21 @@
-import { StructureSchema, ValueType } from '@ephox/boulder';
+import { FieldSchema, StructureSchema, ValueType } from '@ephox/boulder';
 import { Result } from '@ephox/katamari';
 
 import { FormComponentWithLabel, formComponentWithLabelFields, FormComponentWithLabelSpec } from './FormComponent';
 
 export interface ColorInputSpec extends FormComponentWithLabelSpec {
   type: 'colorinput';
+  storageKey?: string;
 }
 
 export interface ColorInput extends FormComponentWithLabel {
   type: 'colorinput';
+  storageKey: string;
 }
 
-const colorInputFields = formComponentWithLabelFields;
+const colorInputFields = formComponentWithLabelFields.concat([
+  FieldSchema.defaultedString('storageKey', 'default')
+]);
 
 export const colorInputSchema = StructureSchema.objOf(colorInputFields);