+++ modules/tinymce/src/themes/silver/main/ts/ui/core/color/ColorCache.ts
@@ -1,16 +1,20 @@
-import { Arr, Type } from '@ephox/katamari';
+import { Arr, Obj, Type } from '@ephox/katamari';
 
+import { Menu } from 'tinymce/core/api/ui/Ui';
 import LocalStorage from 'tinymce/core/api/util/LocalStorage';
-
-const storageName = 'tinymce-custom-colors';
 
 export interface ColorCache {
   readonly add: (key: string) => void;
   readonly state: () => string[];
 }
 
-export const ColorCache = (max: number = 10): ColorCache => {
-  const storageString = LocalStorage.getItem(storageName);
+interface CacheStorage {
+  [index: string]: ColorCache;
+}
+const cacheStorage: CacheStorage = {};
+
+const ColorCache = (storageId: string, max: number = 10): ColorCache => {
+  const storageString = LocalStorage.getItem(storageId);
   const localstorage = Type.isString(storageString) ? JSON.parse(storageString) : [];
 
   const prune = (list: string[]): string[] => {
@@ -33,7 +37,7 @@ export const ColorCache = (max: number = 10): ColorCache => {
       cache.pop();
     }
 
-    LocalStorage.setItem(storageName, JSON.stringify(cache));
+    LocalStorage.setItem(storageId, JSON.stringify(cache));
   };
 
   const remove = (idx: number): void => {
@@ -46,4 +50,40 @@ export const ColorCache = (max: number = 10): ColorCache => {
     add,
     state
   };
+};
+
+const getCacheForId = (id: string): ColorCache =>
+  Obj.get(cacheStorage, id).getOrThunk(() => {
+    const storageId = `tinymce-custom-colors-${id}`;
+    const currentData = LocalStorage.getItem(storageId);
+
+    if (Type.isNullable(currentData)) {
+      const legacyDefault = LocalStorage.getItem('tinymce-custom-colors');
+      LocalStorage.setItem(storageId, Type.isNonNullable(legacyDefault) ? legacyDefault : '[]');
+    }
+
+    const storage = ColorCache(storageId, 10);
+    cacheStorage[id] = storage;
+    return storage;
+  });
+
+const getCurrentColors = (id: string): Menu.ChoiceMenuItemSpec[] =>
+  Arr.map(getCacheForId(id).state(), (color) => ({
+    type: 'choiceitem',
+    text: color,
+    value: color
+  }));
+
+const addColor = (id: string, color: string): void => {
+  getCacheForId(id).add(color);
+};
+
+const clearStoredCaches = (): void => {
+  Arr.each(Obj.keys(cacheStorage), (key) => delete cacheStorage[key]);
+};
+
+export {
+  getCurrentColors,
+  addColor,
+  clearStoredCaches
 };