+++ modules/tinymce/src/themes/silver/test/ts/headless/window/WindowManagerAlertTest.ts
@@ -10,10 +10,10 @@ import * as WindowManager from 'tinymce/themes/silver/ui/dialog/WindowManager';
 import * as TestExtras from '../../module/TestExtras';
 
 describe('headless.tinymce.themes.silver.window.WindowManagerAlertTest', () => {
-  const helpers = TestExtras.bddSetup();
+  const extrasHook = TestExtras.bddSetup();
   let windowManager: WindowManagerImpl;
   before(() => {
-    windowManager = WindowManager.setup(helpers.extras());
+    windowManager = WindowManager.setup(extrasHook.access().extras);
   });
 
   const pTeardown = async () => {
@@ -36,7 +36,7 @@ describe('headless.tinymce.themes.silver.window.WindowManagerAlertTest', () => {
   it('Check the basic structure of the alert dialog', async () => {
     createAlert('The alert dialog loads with the basic structure', Fun.noop);
     await pWaitForDialog();
-    const sink = helpers.sink();
+    const sink = extrasHook.access().getDialogSink();
     Assertions.assertStructure('A basic alert dialog should have these components',
       ApproxStructure.build((s, str, arr) => s.element('div', {
         classes: [ arr.has('mce-silver-sink') ],